+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/impl/ProjectModuleServiceImpl.java
@@ -97,19 +97,19 @@ public class ProjectModuleServiceImpl extends ServiceImpl<ProjectModuleMapper, P
 
     @Override
     @Transactional
-    public boolean remove(Long key) {
+    public boolean remove(BigInteger key) {
         boolean result=removeById(key);
         return result ;
     }
 
     @Override
-    public void removeBatch(Collection<Long> idList) {
+    public void removeBatch(Collection<BigInteger> idList) {
         removeByIds(idList);
     }
 
     @Override
     @Transactional
-    public ProjectModule get(Long key) {
+    public ProjectModule get(BigInteger key) {
         ProjectModule et = getById(key);
         if(et==null){
             et=new ProjectModule();
@@ -170,22 +170,22 @@ public class ProjectModuleServiceImpl extends ServiceImpl<ProjectModuleMapper, P
 
 
 	@Override
-    public List<ProjectModule> selectByParent(Long id) {
+    public List<ProjectModule> selectByParent(BigInteger id) {
         return baseMapper.selectByParent(id);
     }
 
     @Override
-    public void removeByParent(Long id) {
+    public void removeByParent(BigInteger id) {
         this.remove(new QueryWrapper<ProjectModule>().eq("parent",id));
     }
 
 	@Override
-    public List<ProjectModule> selectByRoot(Long id) {
+    public List<ProjectModule> selectByRoot(BigInteger id) {
         return baseMapper.selectByRoot(id);
     }
 
     @Override
-    public void removeByRoot(Long id) {
+    public void removeByRoot(BigInteger id) {
         this.remove(new QueryWrapper<ProjectModule>().eq("root",id));
     }
