+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IBranchService.java
@@ -11,7 +11,6 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -30,9 +29,9 @@ public interface IBranchService extends IService<Branch>{
     void createBatch(List<Branch> list) ;
     boolean update(Branch et) ;
     void updateBatch(List<Branch> list) ;
-    boolean remove(Long key) ;
-    void removeBatch(Collection<Long> idList) ;
-    Branch get(Long key) ;
+    boolean remove(BigInteger key) ;
+    void removeBatch(Collection<BigInteger> idList) ;
+    Branch get(BigInteger key) ;
     Branch getDraft(Branch et) ;
     boolean checkKey(Branch et) ;
     boolean save(Branch et) ;
@@ -40,8 +39,8 @@ public interface IBranchService extends IService<Branch>{
     Branch sort(Branch et) ;
     Page<Branch> searchCurProduct(BranchSearchContext context) ;
     Page<Branch> searchDefault(BranchSearchContext context) ;
-    List<Branch> selectByProduct(Long id) ;
-    void removeByProduct(Long id) ;
+    List<Branch> selectByProduct(BigInteger id) ;
+    void removeByProduct(BigInteger id) ;
     /**
      *自定义查询SQL
      * @param sql  select * from table where id =#{et.param}
