+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/ICompanyService.java
@@ -11,7 +11,6 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -30,9 +29,9 @@ public interface ICompanyService extends IService<Company>{
     void createBatch(List<Company> list) ;
     boolean update(Company et) ;
     void updateBatch(List<Company> list) ;
-    boolean remove(Long key) ;
-    void removeBatch(Collection<Long> idList) ;
-    Company get(Long key) ;
+    boolean remove(BigInteger key) ;
+    void removeBatch(Collection<BigInteger> idList) ;
+    Company get(BigInteger key) ;
     Company getDraft(Company et) ;
     boolean checkKey(Company et) ;
     boolean save(Company et) ;
