+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/service/IRelationService.java
@@ -11,7 +11,6 @@ import java.math.BigInteger;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.scheduling.annotation.Async;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.cache.annotation.CacheEvict;
 
@@ -30,9 +29,9 @@ public interface IRelationService extends IService<Relation>{
     void createBatch(List<Relation> list) ;
     boolean update(Relation et) ;
     void updateBatch(List<Relation> list) ;
-    boolean remove(Long key) ;
-    void removeBatch(Collection<Long> idList) ;
-    Relation get(Long key) ;
+    boolean remove(BigInteger key) ;
+    void removeBatch(Collection<BigInteger> idList) ;
+    Relation get(BigInteger key) ;
     Relation getDraft(Relation et) ;
     boolean checkKey(Relation et) ;
     boolean save(Relation et) ;
